Combined the lung scene
I added all the needed aspects of the reworked placement system to the lung scene, there are some issues to where the mouse can't seem to detect where to move. This is something that can be fix.

The towers are basically done. due to some issues with the spawning tower, ill just cut it for now